// contracts/generated/typescript/models/CollectionMembership.ts
export interface CollectionMembership {
    /**
     * 
     * @type {CollectionMembershipMemberTypeEnum}
     * @memberof CollectionMembership
     */
    memberType: CollectionMembershipMemberTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof CollectionMembership
     */
    pageId: string;
    /**
     * 
     * @type {string}
     * @memberof CollectionMembership
     */
    entityId: string;
    /**
     * 
     * @type {string}
     * @memberof CollectionMembership
     */
    displayTitle: string;
    /**
     * 
     * @type {CollectionMembershipSourceTypeEnum}
     * @memberof CollectionMembership
     */
    sourceType: CollectionMembershipSourceTypeEnum;
    /**
     * 
     * @type {string}
     * @memberof CollectionMembership
     */
    sortKey: string;
    /**
     * 
     * @type {string}
     * @memberof CollectionMembership
     */
    sourceRevisionId: string;
    /**
     * 
     * @type {Date}
     * @memberof CollectionMembership
     */
    createdAt: Date;
}


/**
 * @export
 */
export const CollectionMembershipMemberTypeEnum = {
    Page: 'page',
    Entity: 'entity'
} as const;
export type CollectionMembershipMemberTypeEnum = typeof CollectionMembershipMemberTypeEnum[keyof typeof CollectionMembershipMemberTypeEnum];

/**
 * @export
 */
export const CollectionMembershipSourceTypeEnum = {
    Manual: 'manual',
    Rule: 'rule'
} as const;
export type CollectionMembershipSourceTypeEnum = typeof CollectionMembershipSourceTypeEnum[keyof typeof CollectionMembershipSourceTypeEnum];